adding cumulative time report
draws cumulative graph of the amount of data that has been accessed or modified within the last N days. useful to see how much stale data there is and to target data for moving to a lower tier of storage or for deletion. Can also be used to determine a good size for a burst buffer.